requirements.txt
#requirements.txt

telepot
python-dotenv
requests
beautifulsoup4

selenium
webdriver-manager